Bugfix EK space - Checkbox change - splitted components and created folders with css
