+++ src/Components/Checkbox/Checkbox.tsx
@@ -1,9 +1,9 @@
-import './Components.css'
+import './checkbox.css'
 
 export function Checkbox() {
     return <div className="Checkbox"><input type="checkbox"></input></div>;
 }
 
 export function CheckboxChecked() {
-    return <div className="Checkbox"><input type="checkbox" defaultChecked={true}></input></div>;
+    return <div className="Checkbox checked"><input type="checkbox" defaultChecked={true}></input></div>;
 }
