--- src/Components/Components.css
@@ -1,113 +1,0 @@
-.Product {
-    background: #ffffff;
-    box-shadow: 0px 4px 6px rgba(0, 0, 0, 0.1);
-    border-radius: 5px;
-    height: 40px;
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    align-items: center;
-    padding: 0px 12px;
-    font-size: 16px;
-    line-height: 19px;
-    margin: 6px 0px;
-}
-
-.Checkbox {
-    align-content: center;
-    border: 2px solid var(--primary);
-    border-radius: 50%;
-}
-.checked {
-    background: #f2f2f2;
-    box-shadow: none;
-}
-input {
-    width: 18px;
-    height: 18px;
-    background-color: white;
-    border-radius: 50%;
-    vertical-align: middle;
-    border: none;
-    -webkit-appearance: none;
-    outline: none;
-    cursor: pointer;
-    margin: 2px;
-    padding: 0px;
-    transition: 0.3s ease-in-out;
-}
-
-input:checked {
-    background-color: var(--primary);
-    transition: 0.3s ease-in-out;
-}
-
-.inputComp {
-    margin: 0;
-    border: 1px solid #D3D3D3;
-    height: 40px;
-    font-size: 16px;
-    padding-left: 2vw;
-    border-radius: 4px;
-    align-self: center;
-    width: 88vw;
-}
-
-.inputComp:focus {
-    border: 1px solid #FE6019;
-}
-
-.PrimaryButton {
-    width: 90vw;
-    height: 40px;
-    justify-content: space-between;
-    align-items: center;
-    background: #FE6019;
-    color: white;
-    border-radius: 5px;
-    text-align: center;
-    margin-top: 6px;
-    margin-bottom: 6px;
-}
-
-.SecondaryButton {
-    width: 90vw;
-    height: 40px;
-    justify-content: space-between;
-    align-items: center;
-    border: 1px solid #FE6019;
-    color: #FE6019;
-    box-sizing: border-box;
-    border-radius: 5px;
-    text-align: center;
-    margin-top: 6px;
-    margin-bottom: 6px;
-}
-
-label {
-    margin-bottom: 4px;
-    font-size: 16px;
-    font-weight: 300;
-    color: #292929;
-}
-
-.formInput {
-    display: flex;
-    flex-direction: column;
-    margin-top: 6px;
-    margin-bottom: 6px;
-}
-
-.Logo {
-    filter: drop-shadow(0px 4px 4px rgba(0, 0, 0, 0.2));
-    margin-bottom: 6px;
-    padding-right: 28px;
-} 
-
-.logoPlace{
-    align-items: center;
-    font-size: 24px;
-    line-height: 28px;
-    text-align: center;
-    color: #292929;
-}
