+++ src/Components/Logo/logo.css
@@ -1,0 +1,13 @@
+.Logo {
+    filter: drop-shadow(0px 4px 4px rgba(0, 0, 0, 0.2));
+    margin-bottom: 6px;
+    padding-right: 28px;
+} 
+
+.logoPlace{
+    align-items: center;
+    font-size: 24px;
+    line-height: 28px;
+    text-align: center;
+    color: #292929;
+}
